% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_distribution_age.R
\name{calcul_distribution_age}
\alias{calcul_distribution_age}
\title{Une fonction pour avoir la distribution de l'âge dans un data.frame}
\usage{
calcul_distribution_age(df)
}
\arguments{
\item{df}{Un data frame contenant au minimum la colonne :
\itemize{
\item \code{Date.de.naissance} (indiquant la date de naissance des élus).
}}
}
\value{
Un vecteur numérique contenant les quantiles des âges des élus :
\itemize{
\item 0\% (minimum)
\item 25\% (premier quartile)
\item 50\% (médiane)
\item 75\% (troisième quartile)
\item 100\% (maximum)
}
}
\description{
Cette fonction calcule la distribution des âges des élus en renvoyant
les quantiles (minimum, quartiles et maximum) des âges des élus d'un data frame donné.
}
\examples{
# Exemple avec une liste de data frames représentant plusieurs communes
sapply(df_list, calcul_distribution_age)
}
