% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compter_nb_adjoints.R
\name{compter_nombre_d_adjoints}
\alias{compter_nombre_d_adjoints}
\title{Cette fonction calcule le nombre d'élus ayant la fonction d'adjoint
dans un data frame contenant des informations sur les élus municipaux}
\usage{
compter_nombre_d_adjoints(df)
}
\arguments{
\item{df}{Un data frame contenant au minimum la colonne :
\itemize{
\item \code{Libellé.de.la.fonction} (qui indique la fonction de l'élu).
}}
}
\value{
Un entier représentant le nombre d'élus ayant une fonction contenant le mot "adjoint".
}
\description{
Cette fonction calcule le nombre d'élus ayant la fonction d'adjoint
dans un data frame contenant des informations sur les élus municipaux
}
\examples{
# Exemple avec une liste de data frames représentant plusieurs communes
sapply(df_list, compter_nombre_d_adjoints)

}
