% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_professions.R
\name{plot_code_professions}
\alias{plot_code_professions}
\title{Une fonction pour avoir un graphique du nombre des différents codes professionels}
\usage{
plot_code_professions(df)
}
\arguments{
\item{df}{Un data frame contenant au minimum la colonne :
\itemize{
\item \code{Code.de.la.catégorie.socio.professionnelle} (indiquant la profession de chaque élu).
}}
}
\value{
Un graphique en barres montrant la répartition des élus par code socio-professionnel.
}
\description{
Cette fonction génère un graphique en barres horizontales représentant
le nombre d'élus pour chaque code socio-professionnel.
}
