% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_commune.R
\name{plot_commune}
\alias{plot_commune}
\title{Visualiser la répartition des élus par code professionnel pour une commune}
\usage{
plot_commune(df)
}
\arguments{
\item{df}{Un data frame contenant au minimum les colonnes :
\itemize{
\item \code{Libellé.de.la.commune} (nom de la commune),
\item \code{Libellé.du.département} (nom du département),
\item \code{Code.de.la.catégorie.socio.professionnelle} (code professionnel des élus).
}}
}
\value{
Un graphique en barres représentant la répartition des élus par code professionnel dans la commune.
}
\description{
Cette fonction génère un graphique en barres horizontales illustrant
le nombre d'élus pour chaque code socio-professionnel d'une commune donnée.
}
\examples{
# Exemple avec deux jeux de données représentant des communes
lapply(list(df_Nantes, df_Faverelles), plot_commune)
}
