% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_departement.R
\name{plot_departement}
\alias{plot_departement}
\title{Visualisation des professions des élus par département}
\usage{
plot_departement(df)
}
\arguments{
\item{df}{Un data frame contenant les informations des élus, incluant
les colonnes \code{Libellé.du.département}, \code{Libellé.de.la.commune},
et \code{Code.de.la.catégorie.socio.professionnelle}.}
}
\value{
Un graphique en barres représentant les 10 catégories socio-professionnelles
les plus fréquentes parmi les élus du département.
}
\description{
Cette fonction génère un graphique en barres horizontales illustrant
la répartition des élus selon leur catégorie socio-professionnelle
pour un département donné.
}
\examples{
# Exemple avec un jeu de données spécifique
plot_departement(df_Loire_Atlantique)
plot_departement(df_Gers)
}
