% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_commune.R
\name{summary_commune}
\alias{summary_commune}
\title{Résumé des informations d'une commune}
\usage{
summary_commune(df)
}
\arguments{
\item{x}{Un objet de type \code{data.frame} représentant une commune et contenant
les colonnes \code{Libellé.de.la.commune} ainsi que les informations sur les élus.}
}
\value{
Affiche un résumé des informations de la commune, notamment :
\itemize{
\item Nom de la commune
\item Nombre total d'élus
\item Distribution des âges des élus
\item Élu le plus âgé
}
}
\description{
Cette fonction affiche un résumé des caractéristiques principales
d'une commune donnée, notamment son nom, le nombre d'élus,
la distribution des âges des élus, et l'âge de l'élu le plus âgé.
}
\examples{
# Exemple avec deux jeux de données représentant des communes
summary_commune(df_Nantes)
summary_commune(df_Faverelles)
}
