% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elu_le_plus_age.R
\name{trouver_l_elu_le_plus_age}
\alias{trouver_l_elu_le_plus_age}
\title{Une fonction pour trouver l'élu le plus agé dans un data frame
Cette fonction identifie et affiche l'élu le plus âgé dans un
data frame contenant des informations sur les élus.}
\usage{
trouver_l_elu_le_plus_age(df)
}
\arguments{
\item{df}{Un data frame contenant au minimum les colonnes :
\itemize{
\item \code{Nom.de.l.élu}
\item \code{Prénom.de.l.élu}
\item \code{Date.de.naissance}
}}
}
\value{
Une chaîne de caractères indiquant le nom, prénom et âge de l'élu le plus âgé.
}
\description{
Une fonction pour trouver l'élu le plus agé dans un data frame
Cette fonction identifie et affiche l'élu le plus âgé dans un
data frame contenant des informations sur les élus.
}
\examples{
# Exemple avec une liste de data frames
sapply(df_list, trouver_l_elu_le_plus_age)
}
